Update profile copy
Move the save button to the left and change label because of the context changes.
It will only save the email address now.
Add context for the password reset link.
Change Reset password to Change password, the desire of the user is to change his or her password, not to reset it.