Change loading pattern to dispatch on mount
Before this change we were dispatching an action as soon as the
container component function is called. If this action triggers a
re-render we need to we are basically trying to render again while we
didn't complete our previous render.

React doesn't like this. So it shows us a nice error: "Cannot update
during an existing state transition (such as within `render` or another
component's constructor). Render methods should be a pure function of
props and state; constructor side-effects are an anti-pattern, but can
be moved to `componentWillMount`."

This change moves the dispatching to the `componentDidMount` lifecycle
method. Because we don't have access to dispatch there we abstract this
away to a generic `loadData` prop. This also has the added bonus that
we aren't dispatch every time any other data is changed that re-renders
the container component.